Add loadProjectList method and projects array
Project constructor now require an array to be
stored in as a parameter. Projects are now stored
in projects array in index.js.

Add loadProjectList in DOMHandler.js to load DOM
elements from a list of projects from a
projectsArray into a parentContainer. It also
creates a button to add new projects through a
text input element and confirm/cancel buttons.
It is used in index.js to populate the left nav
bar.

Add basic style in style.css to better visualize
new functionalities.