Add storageHandler.js and save/load functionality
Add storageHandler.js with two methods:

setSavedProjects save an array of project objects
as JSON. Used each time we add or remove a project
or todo.

getSavedProjects return the parsed JSON. Used on
page initial load.

addTodo and removeTodo methods in project.js now
require a projectList as a parameter to enable
saving the list to localStorage on those events.